requirements.txt
Flask==3.0.2
mysql-connector-python==9.4.0